refactor process_reaction
