+++ app.py
@@ -6,7 +6,6 @@ from discord.ext.commands.errors import CommandNotFound
 from pymongo import MongoClient
 #Python
 import asyncio
-from typing import Optional
 from datetime import datetime as dtime
 from datetime import timezone, timedelta
 import os
@@ -139,15 +138,22 @@ async def on_guild_remove(guild):
 
 @bot.event
 async def on_raw_reaction_add(payload):
-    # get reaction from payload
+    
     if not payload.guild_id:
         return
+
+    # ignore if not one of the wanted emotes
+    if str(payload.emoji) not in ('✅', u"\U0001F4C5", u"\U0001F6AB"):
+        return
+
     channel = bot.get_channel(payload.channel_id)
     permissions = channel.permissions_for(channel.guild.me)
     
+    # abort if not able to read messages in the channel of reaction
     if not permissions.read_message_history:
         return
 
+    # get reaction from payload
     try:
         msg = await channel.fetch_message(payload.message_id)
         reaction = discord.utils.get(msg.reactions, emoji=payload.emoji.name)
@@ -162,8 +168,7 @@ async def on_raw_reaction_add(payload):
             if not reaction.me:
                 return
             if msg.embeds:
-                user = bot.get_user(payload.user_id)
-                await member_handler.process_reaction(channel, msg, user, reaction)
+                await process_reaction(channel, msg, reaction)
                     
     except discord.errors.Forbidden:
         logging.info("%s: forbidden on reaction in %s", payload.guild_id, channel.id)
@@ -173,6 +178,148 @@ async def on_raw_reaction_add(payload):
         return
     except discord.errors.DiscordServerError:
         logging.info("%s: Discord Server has some problems", payload.guild_id)
+
+
+async def process_reaction(channel, msg, reaction):
+    emoji = reaction.emoji
+    embed = msg.embeds[0]
+    server_db = database.get_server_db(msg.guild.id)
+    threads_enabled = server_db.get_threads_enabled()
+    success = False
+
+    # always only the id
+    target_member_id = int(embed.title)
+
+    # correct date
+    if emoji == '✅':
+        logging.info("Recognized date correct in %s for user %s.", channel.guild.id, target_member_id)
+        
+        if server_db.get_automatic():
+            membership_date = embed.fields[0].value
+
+            # set membership
+            if await member_handler.set_membership(msg, target_member_id, membership_date, False, msg.author):
+                await asyncio.sleep(0.21)
+                await msg.clear_reactions()
+                await asyncio.sleep(0.21)
+                await msg.add_reaction(emoji='👌')
+                success = True
+
+    # wrong date
+    elif emoji == u"\U0001F4C5":
+        logging.info("Wrong date recognized in %s for user %s.", channel.guild.id, target_member_id)
+
+        success = await handle_wrong_date(channel, msg, reaction, target_member_id)
+
+    # deny option - fake / missing date
+    elif emoji == u"\U0001F6AB":
+        logging.info("Fake or without date in %s for user %s.", channel.guild.id, target_member_id)
+        success = await handle_denied(channel, msg, reaction, embed, target_member_id)
+
+    if success and threads_enabled:
+        log_channel = bot.get_channel(server_db.get_log_channel())
+        embed.clear_fields()
+        embed.set_image(url = discord.Embed.Empty)
+        await log_channel.send(content=None, embed = embed)
+        await channel.edit(archived=True)
+
+async def handle_wrong_date(channel, msg, reaction, target_member_id: int) -> bool:
+    """Process if the date was recognized wrongly
+    
+    Parameters
+    ----------
+    channel:
+        The channel of the proof message
+    msg:
+        The proof message
+    reaction:
+        The reaction that was added
+    target_member_id: int
+        The id of the member whose proof is being processed
+
+    Returns
+    -------
+    bool
+        Whether the process was ended successfully (no abort)
+    """
+
+    m = "Please write the correct date from the screenshot in the format dd/mm/yyyy.\n"
+    m += "Type CANCEL to stop the process."
+    await channel.send(m, reference=msg, mention_author=False)
+
+    user = msg.author
+
+    def check(m):
+        return m.author == user and m.channel == channel
+
+    date_msg = await bot.wait_for('message', check=check)
+
+    if date_msg.content.lower() != "cancel" and await member_handler.set_membership(msg, target_member_id, date_msg.content, False, user):
+        await msg.clear_reactions()
+        await asyncio.sleep(0.21)
+        await msg.add_reaction(emoji='👍')
+        return True
+    else:
+        logging.info("Canceled reaction by user %s in %s.", user.id, channel.guild.id)
+        await reaction.remove(user)
+        await asyncio.sleep(0.21)
+        await channel.send("Stopped the process and removed reaction.")
+        return False
+
+
+
+async def handle_denied(channel, msg, reaction, embed, target_member_id: int) -> bool:
+    """Process if the proof is denied
+    
+    Parameters
+    ----------
+    channel:
+        The channel of the proof message
+    msg:
+        The proof message
+    reaction:
+        The reaction that was added
+    embed:
+        The edited embed
+    target_member_id: int
+        The id of the member whose proof is being processed
+
+    Returns
+    -------
+    bool
+        Whether the process was ended successfully (no abort)
+    """
+
+    m = "Please write a message that will be sent to the User."
+    m += "Type CANCEL to stop the process."
+    await channel.send(m, reference=msg, mention_author=False)
+
+    user = msg.author
+
+    def check(m):
+        return m.author == user and m.channel == channel
+
+    text_msg = await bot.wait_for('message', check=check)
+    if text_msg.content.lower() != "cancel":
+        target_member = bot.get_user(target_member_id)
+        await target_member.send("{} server:\n{}".format(Utility.get_vtuber(msg.guild.id), text_msg.content))
+        await channel.send("Message was sent to {}.".format(target_member.mention), reference=text_msg, mention_author=False)
+
+        if database.get_server_db(msg.guild.id).get_automatic():
+            await member_handler.del_membership(msg, target_member_id, None, False, False)
+            # set embed
+        embed.description = "**DENIED**\nUser: {}\nBy: {}".format(target_member.mention, user)
+        await msg.edit(content = msg.content, embed = embed)
+        await asyncio.sleep(0.21)
+        await msg.clear_reactions()
+        await msg.add_reaction(emoji='👎')
+        return True
+    else:
+        logging.info("Canceled reaction by user %s in %s.", user.id, channel.guild.id)
+        await reaction.remove(user)
+        await channel.send("Stopped the process and removed reaction.")
+        return False
+
 
 def dm_or_test_only():
     def predicate(ctx):
