+++ membership_handling.py
@@ -533,104 +533,3 @@ class MembershipHandler:
                 await asyncio.sleep(10) # check all 10 seconds
             except Exception:
                 logging.exception("Catched error in deque")
-
-    async def process_reaction(self, channel, msg, user, reaction):
-        emoji = reaction.emoji
-        embed = msg.embeds[0]
-        automatic_role = self.db.get_server_db(msg.guild.id).get_automatic()
-        threads_enabled = self.db.get_server_db(msg.guild.id).get_threads_enabled()
-        bot = self.bot
-        success = False
-
-        
-        # always only the id
-        target_member_id = int(embed.title)
-
-        # correct date
-        if emoji == '✅':
-            logging.info("Recognized date correct in %s for user %s.", channel.guild.id, target_member_id)
-            
-            if not automatic_role:
-                membership_date = embed.fields[0].value
-
-                # set membership
-                if await self.set_membership(msg, target_member_id, membership_date, False, user):
-                    await asyncio.sleep(0.21)
-                    await msg.clear_reactions()
-                    await asyncio.sleep(0.21)
-                    await msg.add_reaction(emoji='👌')
-                    success = True
-
-        # wrong date
-        elif emoji == u"\U0001F4C5":
-            logging.info("Wrong date recognized in %s for user %s.", channel.guild.id, target_member_id)
-
-            success = await self.handle_wrong_date(channel, msg, user, reaction, bot, target_member_id)
-
-        # deny option - fake / missing date
-        elif emoji == u"\U0001F6AB":
-            logging.info("Fake or without date in %s for user %s.", channel.guild.id, target_member_id)
-            success = await self.handle_denied(channel, msg, user, reaction, embed, automatic_role, bot, target_member_id)
-
-        if success and threads_enabled:
-            log_channel = bot.get_channel(self.db.get_server_db(msg.guild.id).get_log_channel())
-            embed.clear_fields()
-            embed.set_image(url = discord.Embed.Empty)
-            await log_channel.send(content=None, embed = embed)
-            await channel.edit(archived=True)
-
-
-
-
-    async def handle_wrong_date(self, channel, msg, user, reaction, bot, target_member_id) -> bool:
-        m = "Please write the correct date from the screenshot in the format dd/mm/yyyy.\n"
-        m += "Type CANCEL to stop the process."
-        await channel.send(m, reference=msg, mention_author=False)
-
-        def check(m):
-            return m.author == user and m.channel == channel
-
-        date_msg = await bot.wait_for('message', check=check)
-
-        if date_msg.content.lower() != "cancel" and await self.set_membership(msg, target_member_id, date_msg.content, False, user):
-            await msg.clear_reactions()
-            await asyncio.sleep(0.21)
-            await msg.add_reaction(emoji='👍')
-            return True
-        else:
-            logging.info("Canceled reaction by user %s in %s.", user.id, channel.guild.id)
-            await reaction.remove(user)
-            await asyncio.sleep(0.21)
-            await channel.send("Stopped the process and removed reaction.")
-            return False
-
-
-
-    async def handle_denied(self, channel, msg, user, reaction, embed, automatic_role, bot, target_member_id) -> bool:
-        m = "Please write a message that will be sent to the User."
-        m += "Type CANCEL to stop the process."
-        await channel.send(m, reference=msg, mention_author=False)
-
-        def check(m):
-            return m.author == user and m.channel == channel
-
-        text_msg = await bot.wait_for('message', check=check)
-        if text_msg.content.lower() != "cancel":
-            target_member = bot.get_user(target_member_id)
-            await target_member.send("{} server:\n{}".format(Utility.get_vtuber(msg.guild.id), text_msg.content))
-            await channel.send("Message was sent to {}.".format(target_member.mention), reference=text_msg, mention_author=False)
-
-            if automatic_role:
-                await self.del_membership(msg, target_member_id, None, False, False)
-                # set embed
-            embed.description = "**DENIED**\nUser: {}\nBy: {}".format(target_member.mention, user)
-            await msg.edit(content = msg.content, embed = embed)
-            await asyncio.sleep(0.21)
-            await msg.clear_reactions()
-            await msg.add_reaction(emoji='👎')
-            return True
-        else:
-            logging.info("Canceled reaction by user %s in %s.", user.id, channel.guild.id)
-            await reaction.remove(user)
-            await channel.send("Stopped the process and removed reaction.")
-            return False
